+++ inc/core/patterns/bcgov-accordion-with-media-text.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace Bcgov\Theme\Block;
+
+/**
+ * General Accordion with Media Text block pattern.
+ *
+ * @since 1.0.4
+ *
+ * @package Bcgov/Theme/Block
+ */
+return [
+	'title'      => __( 'Accordion with Media Text', 'bcgov_blocks_theme' ),
+	'categories' => [ 'bcgov-blocks-theme-general' ],
+	'content'    => '<!-- wp:bcgov-block-theme/collapse {"collapseId":"collapse-container-38c96374-0bfe-4349-bcdb-b67d50c6ba66","align":"wide"} -->
+	<div id="collapse-container-38c96374-0bfe-4349-bcdb-b67d50c6ba66" class="wp-block-bcgov-block-theme-collapse alignwide"><div class="collapse-container-nav"><span><button data-target="#collapse-container-38c96374-0bfe-4349-bcdb-b67d50c6ba66" class="collapse-expand-all">Expand all</button></span><span><button data-target="#collapse-container-38c96374-0bfe-4349-bcdb-b67d50c6ba66" class="collapse-collapse-all">Collapse all</button></span></div><!-- wp:bcgov-block-theme/collapse-item {"title":"Lorem ipsum dolar sit amut","itemId":"collapse-item-b31fb960-947e-4c39-b171-193dfad97f7a","headingId":"heading-b31fb960-947e-4c39-b171-193dfad97f7a"} -->
+	<div class="wp-block-bcgov-block-theme-collapse-item"><div class="collapse-header" id="heading-b31fb960-947e-4c39-b171-193dfad97f7a"><h3><button data-toggle="collapse" data-target="#collapse-item-b31fb960-947e-4c39-b171-193dfad97f7a" aria-expanded="false" aria-controls="collapse-item-b31fb960-947e-4c39-b171-193dfad97f7a" class="collapsed"><span class="collapse-title">Lorem ipsum dolar sit amut</span></button></h3></div><div class="collapse collapse-container" id="collapse-item-b31fb960-947e-4c39-b171-193dfad97f7a"><div class="collapse-body"><!-- wp:media-text {"mediaId":1752,"mediaLink":"none","mediaType":"image","mediaWidth":35,"verticalAlignment":"center","style":{"spacing":{"padding":{"top":"0","right":"0","bottom":"0","left":"0"},"margin":{"top":"0","right":"0","bottom":"0","left":"0"}}},"className":"in-page-full-width-constrained-section floating is-style-no-clip"} -->
+	<div class="wp-block-media-text alignwide is-stacked-on-mobile is-vertically-aligned-center in-page-full-width-constrained-section floating is-style-no-clip" style="margin-top:0;margin-right:0;margin-bottom:0;margin-left:0;padding-top:0;padding-right:0;padding-bottom:0;padding-left:0;grid-template-columns:35% auto"><figure class="wp-block-media-text__media"><img src="' . esc_url( get_template_directory_uri() ) . '/assets/images/sites/cleanbc/cleanbc-fish-placeholder.jpg" alt="" class="wp-image-1752 size-full"/></figure><div class="wp-block-media-text__content"><!-- wp:group -->
+	<div class="wp-block-group"><!-- wp:heading {"level":4,"textColor":"secondary-brand"} -->
+	<h4 class="has-secondary-brand-color has-text-color"><strong>Vivamus sagittis lacus vel augue laoreet rutrum.</strong></h4>
+	<!-- /wp:heading -->
+	
+	<!-- wp:heading {"level":5,"textColor":"secondary-brand"} -->
+	<h5 class="has-secondary-brand-color has-text-color">Ipsum Inceptos Condimentum.</h5>
+	<!-- /wp:heading -->
+	
+	<!-- wp:paragraph {"textColor":"primary-brand"} -->
+	<p class="has-primary-brand-color has-text-color">Donec id elit non mi porta gravida at eget metus. Maecenas faucibus mollis interdum. Donec id elit non mi porta gravida at eget metus. Sed posuere consectetur est at lobortis.</p>
+	<!-- /wp:paragraph -->
+	
+	<!-- wp:buttons {"className":"is-bcgov"} -->
+	<div class="wp-block-buttons is-bcgov"><!-- wp:button {"className":"is-style-underline","iconsList":[{"label":"None","value":""},{"label":"Default","value":"default"},{"label":"Home","value":"home"},{"label":"Auto","value":"auto"},{"label":"Book","value":"book"},{"label":"Calendar","value":"calendar"},{"label":"Case","value":"case"},{"label":"Changes","value":"changes"},{"label":"Federal","value":"federal"},{"label":"Gear","value":"gear"},{"label":"Group","value":"group"},{"label":"Leaf","value":"leaf"},{"label":"Map","value":"map"},{"label":"Tag","value":"tag"}]} -->
+	<div class="wp-block-button is-style-underline"><a class="wp-block-button__link wp-element-button">Learn more about vivamus sagittis</a></div>
+	<!-- /wp:button --></div>
+	<!-- /wp:buttons --></div>
+	<!-- /wp:group --></div></div>
+	<!-- /wp:media-text --><div class="collapse-close"><a data-toggle="collapse" data-target="#collapse-item-b31fb960-947e-4c39-b171-193dfad97f7a" href="#collapse-item-b31fb960-947e-4c39-b171-193dfad97f7a" role="button" aria-expanded="true" aria-controls="collapse-item-b31fb960-947e-4c39-b171-193dfad97f7a">Collapse</a></div></div></div></div>
+	<!-- /wp:bcgov-block-theme/collapse-item -->
+	
+	<!-- wp:bcgov-block-theme/collapse-item {"title":"Sollicitudin Vestibulum Dolor Inceptos","itemId":"collapse-item-235df830-0ba4-40b3-ab37-1097144d55ae","headingId":"heading-235df830-0ba4-40b3-ab37-1097144d55ae"} -->
+	<div class="wp-block-bcgov-block-theme-collapse-item"><div class="collapse-header" id="heading-235df830-0ba4-40b3-ab37-1097144d55ae"><h3><button data-toggle="collapse" data-target="#collapse-item-235df830-0ba4-40b3-ab37-1097144d55ae" aria-expanded="false" aria-controls="collapse-item-235df830-0ba4-40b3-ab37-1097144d55ae" class="collapsed"><span class="collapse-title">Sollicitudin Vestibulum Dolor Inceptos</span></button></h3></div><div class="collapse collapse-container" id="collapse-item-235df830-0ba4-40b3-ab37-1097144d55ae"><div class="collapse-body"><!-- wp:media-text {"mediaPosition":"right","mediaId":1752,"mediaLink":"none","mediaType":"image","mediaWidth":35,"verticalAlignment":"center","imageFill":false,"textColor":"background","className":"in-page-full-width-constrained-section floating is-style-no-clip"} -->
+	<div class="wp-block-media-text alignwide has-media-on-the-right is-stacked-on-mobile is-vertically-aligned-center in-page-full-width-constrained-section floating is-style-no-clip has-background-color has-text-color" style="grid-template-columns:auto 35%"><div class="wp-block-media-text__content"><!-- wp:group {"layout":{"type":"default"}} -->
+	<div class="wp-block-group"><!-- wp:heading {"level":4,"textColor":"secondary-brand"} -->
+	<h4 class="has-secondary-brand-color has-text-color"><strong>Vivamus sagittis lacus vel augue laoreet rutrum.</strong></h4>
+	<!-- /wp:heading -->
+	
+	<!-- wp:heading {"level":5,"textColor":"secondary-brand"} -->
+	<h5 class="has-secondary-brand-color has-text-color">Ipsum Inceptos Condimentum.</h5>
+	<!-- /wp:heading -->
+	
+	<!-- wp:paragraph {"textColor":"primary-brand"} -->
+	<p class="has-primary-brand-color has-text-color">Donec id elit non mi porta gravida at eget metus. Maecenas faucibus mollis interdum. Donec id elit non mi porta gravida at eget metus. Sed posuere consectetur est at lobortis.</p>
+	<!-- /wp:paragraph -->
+	
+	<!-- wp:buttons {"className":"is-bcgov"} -->
+	<div class="wp-block-buttons is-bcgov"><!-- wp:button {"className":"is-style-underline","iconsList":[{"label":"None","value":""},{"label":"Default","value":"default"},{"label":"Home","value":"home"},{"label":"Auto","value":"auto"},{"label":"Book","value":"book"},{"label":"Calendar","value":"calendar"},{"label":"Case","value":"case"},{"label":"Changes","value":"changes"},{"label":"Federal","value":"federal"},{"label":"Gear","value":"gear"},{"label":"Group","value":"group"},{"label":"Leaf","value":"leaf"},{"label":"Map","value":"map"},{"label":"Tag","value":"tag"}]} -->
+	<div class="wp-block-button is-style-underline"><a class="wp-block-button__link wp-element-button">Learn more about vivamus sagittis</a></div>
+	<!-- /wp:button --></div>
+	<!-- /wp:buttons --></div>
+	<!-- /wp:group --></div><figure class="wp-block-media-text__media"><img src="' . esc_url( get_template_directory_uri() ) . '/assets/images/sites/cleanbc/cleanbc-fish-placeholder.jpg" alt="" class="wp-image-1752 size-full"/></figure></div>
+	<!-- /wp:media-text --><div class="collapse-close"><a data-toggle="collapse" data-target="#collapse-item-235df830-0ba4-40b3-ab37-1097144d55ae" href="#collapse-item-235df830-0ba4-40b3-ab37-1097144d55ae" role="button" aria-expanded="true" aria-controls="collapse-item-235df830-0ba4-40b3-ab37-1097144d55ae">Collapse</a></div></div></div></div>
+	<!-- /wp:bcgov-block-theme/collapse-item --></div>
+	<!-- /wp:bcgov-block-theme/collapse -->',
+];